fix: splitting code on hackathon card
