+++ .gitignore
@@ -29,6 +29,7 @@ yarn-error.log*
 # Env files:
 .env
 .env*.local
+.env.*
 
 # Vercel:
 .vercel
